move createCustomEvent to utils
